Add reference.md
The format sucks, but I don't want to set up a website or something like
that. Also because of the header-only nature the header is super messy
and full of stuff a user should not care about. If anyone ever uses this
library, this has to be replaced, but no one will, so it's okay.